// apps/frontend/panel-react/src/app/api/types.ts
export interface IFetchRes<T = unknown> {
  data: T | null;
  error: number | null;
  newCookiesRaw?: string[] | null;
}

export interface IFetchArgs {
  url: string;
  method?: "GET" | "POST" | "PUT" | "PATCH" | "DELETE";
  body?: unknown;
  params?: Record<string, unknown>;
  credentials?: RequestCredentials;
  headers?: HeadersInit;
  mode?: RequestMode;
}